refactor(browser): extract CDPBrowserBackend base class from ElectronBackend
